fix(bakcend): ディレクトリ構造を修正
